requirements.txt
streamlit>=1.22.0
pandas
numpy
faiss-cpu
sentence-transformers
requests 